pyproject.toml
[project]
name = "excel2sqlite-cli"
version = "0.1.0"
description = "Command line tool to convert Excel Workbook to SQLite3 DB using YAML config"
authors = [
    { name="Your Name", email="your.email@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.0",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0"
]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"